+++ src/rust/engine/remote_provider/src/lib.rs
@@ -29,7 +29,9 @@ use std::sync::Arc;
 
 // Re-export these so that consumers don't have to know about the exact arrangement of underlying
 // crates.
-pub use remote_provider_traits::{ByteStoreProvider, LoadDestination, RemoteOptions};
+pub use remote_provider_traits::{
+  ActionCacheProvider, ByteStoreProvider, LoadDestination, RemoteOptions,
+};
 
 pub const REAPI_ADDRESS_SCHEMAS: [&str; 4] = ["grpc://", "grpcs://", "http://", "https://"];
